requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cramjam==2.8.3
cryptography==42.0.8
daphne==4.1.2
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
kombu==5.3.7
Markdown==3.6
mongoengine==0.28.2
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pykerberos==1.2.4
pymongo==4.7.3
pyOpenSSL==24.1.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-snappy==0.7.1
pyzbar==0.1.9
qrcode==7.4.2
redis==5.0.8
requests==2.32.3
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.4.post2
